+++ NEWS.md
@@ -1,3 +1,5 @@
+# workflows (development version)
+
 # workflows 1.1.3
 
 * The workflows methods for `generics::tune_args()` and `generics::tunable()`