+++ NEWS.md
@@ -1,6 +1,6 @@
 # workflows (development version)
 
-* Each of the `pull_*()` functions soft-deprecated in workflows v0.2.3 now warn on every usage.
+* Each of the `pull_*()` functions soft-deprecated in workflows v0.2.3 now warn on every usage. 
 
 # workflows 1.1.3